refactor(cli): extract the --json document builders into documents.py
The six per-command `--json` builders and their schema constants lived as
private functions in cli.py, where they had grown to ~314 lines of the
dispatch module and were reachable only through argparse. The contract they
serve (machine.py) is not a CLI feature: the planned web backend needs to
call these directly rather than shell out to the CLI and parse its stdout.

Move them verbatim to a new documents.py — the library-level read-model
projection layer, domain object in, contract document dict out. This is the
split probe.py and diagnostics.py already make ("one finding, two render
targets"), generalized to the commands whose document is a dict rather than
a rendered string.

Pure relocation: zero behavior change, zero schema change. The moved bodies
are byte-identical to their previous definitions, and the existing tests —
which pin doc["schema_version"] == cli.X_SCHEMA_VERSION — pass untouched,
which is what proves it. cli.py re-imports the names so those references
keep resolving.

_run_token_totals moves too: _status_document calls it, and leaving it in
cli.py would have made documents.py import the CLI. cmd_status still uses
it for the text rendering, via the same re-import.

Verified: full suite green with zero test edits (2577 passed), and
`status --json` / `list --json` byte-identical before vs after on a fixture
run covering all four terminal statuses, a weighted-token snapshot and a
deferred task.
